+++ src/auth/auth.controller.ts
@@ -116,9 +116,9 @@ export class AuthController {
       const isValid = await this.authService.verifyOtp(userInfo, otp.otp);
       if (isValid) {
         if (userInfo.two_fa === true) {
-          new HttpException("너 잘못했잖아", HttpStatus.BAD_REQUEST);
+          throw new HttpException("너 잘못했잖아", HttpStatus.BAD_REQUEST);
         }
-        await this.authService.setOtpSecret(userInfo);
+        //await this.authService.setOtpSecret(userInfo);
       } else {
         throw new HttpException(
           "OTP 코드가 일치하지 않습니다.",
