+++ src/user/user.otp.service.ts
@@ -3,6 +3,7 @@ import { InjectRepository } from "@nestjs/typeorm";
 import { User } from "./entity/user.entity";
 import { In, Like, Repository } from "typeorm";
 import { UserOtpSecret } from "./entity/user.otp.entity";
+import Redis from "ioredis";
 
 @Injectable()
 export class UserOtpService {
@@ -10,6 +11,7 @@ export class UserOtpService {
   constructor(
     @InjectRepository(UserOtpSecret)
     private userOtpRepositoty: Repository<UserOtpSecret>,
+    private redisService: Redis,
   ) {}
 
   async resetSecret(user: User): Promise<void> {
@@ -35,11 +37,49 @@ export class UserOtpService {
         where: { user_id: user.id },
       });
 
+      return userOtpSecret;
+    } catch (error) {
+      throw error;
+    }
+  }
+
+  async getOtpSecret(user: User): Promise<UserOtpSecret> {
+    try {
+      const userOtpSecret = await this.userOtpRepositoty.findOne({
+        where: { user_id: user.id },
+      });
+
       if (!userOtpSecret) {
         throw new Error("Secret key not found for the user");
       }
 
       return userOtpSecret;
+    } catch (error) {
+      throw error;
+    }
+  }
+
+  async setOtpSecret(user: User): Promise<void> {
+    try {
+      const userOtpInfo = await this.findOneUserOtpInfo(user);
+
+      const storedSecret = await this.redisService.get(`OTP|${user.id}`);
+
+      if (userOtpInfo) {
+        await this.userOtpRepositoty.update(
+          { user_id: user.id },
+          { secret: storedSecret, updated_at: new Date() },
+        );
+        await this.redisService.del(`OTP|${user.id}`);
+      } else {
+        const newUserOtpSecret = {
+          user_id: user.id,
+          secret: storedSecret,
+          updated_at: new Date(),
+        };
+        await this.userOtpRepositoty.save(newUserOtpSecret);
+        await this.redisService.del(`OTP|${user.id}`);
+      }
     } catch (error) {
       throw error;
     }
